Fix --disable-monitor to work with the parenting of Postgres model.
In the current version the pg_autoctl process is the parent process of the
Postgres main process. Except when using --disable-monitor, which makes no
sense, so here we fix that.

It is now possible to control and run a Postgres process under a pg_autoctl
process without using a monitor, as in the following example:

  $ pg_autoctl create postgres --auth trust --no-ssl --disable-monitor --run
  $ PG_AUTOCTL_DEBUG=1 pg_autoctl do fsm assign single

The `pg_autoctl do fsm assign` command now edits the keeper state file, and
the FSM transition that is necessary is then implemented in the keeper loop
as usual.